backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0

# Database
asyncpg==0.29.0
sqlalchemy==1.4.50
databases[postgresql]==0.8.0
alembic==1.12.1  # For database migrations

# HTTP and API clients
httpx==0.25.2
requests==2.31.0
delta-rest-client==0.1.5  # Pinning version for stability

# Security
cryptography==41.0.7
bcrypt==4.0.1  # For password hashing
python-jose[cryptography]==3.3.0  # For JWT tokens
pyjwt==2.8.0  # JWT implementation

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging and monitoring
python-json-logger==2.0.7